Use firebase listeners to redirect users to pages
For example, on lobby start make countdown and redirect to role-reveal
On player death display on role reveal page
On new game start, redirect back to lobby

Right now most of it is functional and does not look visually great.
Lots of loose ends to tie up and refactor and stuff. But, for the most
part, I have an MVP rn